Moved all large listeners out of the GUI class
