+++ src/game/ImagePanel.java
@@ -1,0 +1,58 @@
+package game;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class ImagePanel extends JPanel {
+
+    /**
+     *
+     */
+    private static final long serialVersionUID = -7315460075240330922L;
+    private Image img;
+    private int row;
+    private int column;
+
+    public ImagePanel(String img) {
+        this(new ImageIcon(img).getImage());
+    }
+
+    public ImagePanel(Image img) {
+        this.img = img;
+        Dimension size = new Dimension(img.getWidth(null), img.getHeight(null));
+        setPreferredSize(size);
+        setMinimumSize(size);
+        setMaximumSize(size);
+        setSize(size);
+        setLayout(null);
+    }
+
+    public void paintComponent(Graphics g) {
+        g.drawImage(img, 0, 0, null);
+    }
+
+    public void setRow(int row) {
+        this.row = row;
+    }
+
+    public int getRow() {
+        return this.row;
+    }
+
+    public void setColumn(int column) {
+        this.column = column;
+    }
+
+    public int getColumn() {
+        return this.column;
+    }
+
+    public int getPixelX() {
+        return this.column * 80 + 10;
+    }
+
+    public int getPixelY() {
+        return this.row * 80 + 10;
+    }
+
+}
